feat: implementar regras customizadas

+++ src/TresCamadas.Business/Services/FornecedorService.cs
@@ -19,7 +19,11 @@ public class FornecedorService : BaseService, IFornecedorService
             !ExecutarValidacao(new EnderecoValidation(), fornecedor.Endereco)) 
             return;
 
-        // Validar se ja nao existe outro fornecedor com o mesmo doc.
+        if (_fornecedorRepository.Buscar(f => f.Documento == fornecedor.Documento).Result.Any())
+        {
+            Notificar("Já existe um fornecedor com este documento infomado.");
+            return;
+        }
 
         await _fornecedorRepository.Adicionar(fornecedor);
     }
@@ -29,11 +33,38 @@ public class FornecedorService : BaseService, IFornecedorService
         if (!ExecutarValidacao(new FornecedorValidation(), fornecedor))
             return;
 
+        if (_fornecedorRepository.Buscar(f => f.Documento == fornecedor.Documento && f.Id != fornecedor.Id).Result.Any())
+        {
+            Notificar("Já existe um fornecedor com este documento infomado.");
+            return;
+        }
+
         await _fornecedorRepository.Atualizar(fornecedor);
     }
 
     public async Task Remover(Guid id)
     {
+        var fornecedor = await _fornecedorRepository.ObterFornecedorProdutosEndereco(id);
+
+        if (fornecedor == null)
+        {
+            Notificar("Fornecedor não existe!");
+            return;
+        }
+
+        if (fornecedor.Produtos.Any())
+        {
+            Notificar("O fornecedor possui produtos cadastrados!");
+            return;
+        }
+
+        var endereco = await _fornecedorRepository.ObterEnderecoPorFornecedor(id);
+
+        if (endereco != null)
+        {
+            await _fornecedorRepository.RemoverEnderecoFornecedor(endereco);
+        }
+
         await _fornecedorRepository.Remover(id);
     }
